Fixing admin and user layouts (exclude second header). Adding component styles. Delete empty components directory. Create UserCards component instead UserTables. Working AdminSidebar.
