+++ app/src/components/AdminSidebar/AdminSidebar.js
@@ -1,31 +1,16 @@
 import React, { PropTypes } from 'react';
-import { Link } from 'react-router-dom';
+import { NavSidebar } from 'Components';
 import AdminSidebarData from './AdminSidebarData';
+import './AdminSidebar.scss';
 
-const AdminSidebar = props => {
-  const links = props.data;
+const AdminSidebar = () => {
+  const tabs = AdminSidebarData.tabs
+
   return (
-    <nav className='admin-sidebar'>
-      <ul className='admin-sidebar-tab'>
-        {links.map((link, i) => {
-          return (
-            <li key={i} className='admin-sidebar-item'>
-              <Link to={links.url} className='admin-sidebar-link'>
-                {links.title}
-              </Link>
-            </li>
-          )
-        }
-        )}
-      </ul>
-    </nav>
+    <div className='admin-sidebar'>
+      <NavSidebar data={tabs} />
+    </div>
   )
 }
-
-AdminSidebar.propTypes = {
-  title: PropTypes.string,
-  url: PropTypes.string,
-  key: PropTypes.number
-};
 
 export default AdminSidebar;
