+++ app/src/components/AdminSidebar/AdminSidebar.scss
@@ -1,0 +1,3 @@
+.admin-sidebar {
+  
+}
