+++ app/src/layouts/AdminLayout.js
@@ -1,13 +1,8 @@
 import React from 'react';
 
-import { Header } from 'Components';
-
 const AdminLayout = props => (
-  <div className='login-layout'>
-    <Header />
-    <div className='layout-content'>
-      {props.children}
-    </div>
+  <div className='admin-layout'>
+    {props.children}
   </div>
 )
