+++ app/src/layouts/UserLayout.js
@@ -1,13 +1,8 @@
 import React from 'react';
 
-import { Header } from 'Components';
-
 const UserLayout = props => (
-  <div className='login-layout'>
-    <Header />
-    <div className='layout-content'>
-      {props.children}
-    </div>
+  <div className='user-layout'>
+    {props.children}
   </div>
 )
