+++ app/src/pages/Cabinet/AdminCabinet.scss
@@ -1,0 +1,4 @@
+.admin-cabinet {
+  height: 700px;
+  background-color: green;
+}
