+++ app/src/pages/Cabinet/UserCabinet.scss
@@ -1,0 +1,4 @@
+.user-cabinet {
+  height: 700px;
+  background-color: gray;
+}
